apps/workflow/requirements.txt
pyaml==21.10.1
requests==2.28.2
macaddress==2.0.2
scapy==2.5.0
boto3==1.17.105